Map view
The map view is in. There is no key present for the mapview SDK as I am
not sure how to do this and keep it out of the repo. What will most
likely happen is that one will register their own key and then keep it
in a non versioned file and call that. But I'll have to figure it out.