Heart rate detection working.
The heart rate detection is now working, it's not 100% accurate but
considering it's coming from a phone its pretty damn good.

I have also added in the camera detection code to give people a nice
error if they for some reason manage to get the app when their phone
won't work with it.